LAB2-reinforcement_learning/requirements.txt
cvxopt
image
ipython
ipythonblocks
ipywidgets
jupyter
matplotlib
networkx
numpy
pandas
pillow
scipy
sortedcontainers